fix(sandbox): add remove_dir/remove_file permissions to sandbox
When an extraction directory is empty, unblob will try to delete it.

This can lead to PermissionError due to insufficient permissions within
the sandbox. The same can happen when deleting files.

We therefore allow unblob to delete directories and files from the
extraction directory with the remove_dir and remove_file permissions.